+++ packages/gov-ui-kit/src/core/components/forms/inputFileAvatar/index.tsx
@@ -1,2 +1,2 @@
 export { InputFileAvatar } from './inputFileAvatar';
-export { InputFileAvatarError, type IInputFileAvatarProps } from './inputFileAvatar.api';
+export { InputFileAvatarError, type IInputFileAvatarProps, type IInputFileAvatarValue } from './inputFileAvatar.api';