+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsList/proposalActionUpdateMetadata/proposalActionUpdateMetadata.testUtils.ts
@@ -8,13 +8,13 @@ export const generateProposalActionUpdateMetadata = (
     ...generateProposalAction(),
     type: ProposalActionType.UPDATE_METADATA,
     existingMetadata: {
-        logo: 'https://i.pravatar.cc/300',
+        avatar: 'https://i.pravatar.cc/300',
         name: 'Old name',
         description: 'Existing DAO description.',
         links: [],
     },
     proposedMetadata: {
-        logo: 'https://i.pravatar.cc/300',
+        avatar: 'https://i.pravatar.cc/300',
         name: 'New name',
         description: 'Proposed DAO description.',
         links: [],